Added basic BFS. Added mouse button event in RunGame to run generatePath using mouse position

+++ DynamicFlowField/Building.cpp
@@ -16,10 +16,10 @@ Building::Building(int size, Toolbox::BuildingType type, sf::Vector2i pos) :
 	float scaleY = mSprite.getTexture()->getSize().y / sizeY;
 
 	// Reset sprite size to fit block
-	mSprite.scale(sf::Vector2f(1/scaleX, 1/scaleY));
+	mSprite.scale(sf::Vector2f(1 / scaleX, 1 / scaleY));
 
 	// Scale sprite to accomodate building size 
-	mSprite.scale(sf::Vector2f(size, size));
+	mSprite.scale(sf::Vector2f((float)size, (float)size));
 	mSprite.setPosition(sf::Vector2f((float)mPosition.x * sizeX, (float)mPosition.y * sizeY));
 }
 
@@ -35,32 +35,6 @@ void Building::update()
 void Building::render(sf::RenderWindow& window)
 {
 	window.draw(mSprite);
-	//for (int i = 0; i < mSize; i++)
-	//{
-	//	for (int j = 0; j < mSize; j++)
-	//	{
-	//		sf::Vector2f realPos = sf::Vector2f((mPosition.x + i) * sizeX, (mPosition.y + j) * sizeY);
-
-	//		sf::RectangleShape rect(sf::Vector2f(sizeX, sizeY));
-	//		rect.setPosition(realPos);
-	//		rect.setOutlineColor(sf::Color::Black);
-	//		rect.setOutlineThickness(1.f);
-
-	//		switch (mType)
-	//		{
-	//		case Toolbox::OFFENSIVE:
-	//			rect.setFillColor(sf::Color::Yellow);
-	//			break;
-	//		case Toolbox::DEFENSIVE:
-	//			rect.setFillColor(sf::Color(50, 50, 50)); // Gray
-	//			break;
-	//		default:
-	//			break;
-	//		}
-
-	//		window.draw(rect);
-	//	}
-	//}
 }
 
 void Building::kill()
